アニメーションとモデルの更新 壁の透明化

+++ Prototype/Element/Assets/Script/UI/UIScript.cs
@@ -17,7 +17,7 @@ public class UIScript : MonoBehaviour {
     void Update()
     {
         string elementType = MyPlayer.GetElementType().ToString();
-        string element = MyPlayer.GetElementValue().ToString();
+        string element = ((int)MyPlayer.GetElementValue()).ToString();
         string maxElement = MyPlayer.GetMaxElementValue().ToString();
         ElementValueText.text = "Element:" + elementType + "-" + element + "/" + maxElement;
